E-commerce
An online furniture shop that allows users to check for various furniture available at the online store and purchase online. The project consists of list of furniture products displayed in various models and designs. The user may browse through these products as per categories. If the user likes a product he may add it to his shopping cart. Now he may pay through a credit card or cash on delivery.